lab 1
laboratory work in C language with assembler insertion. The program receives a string of three characters as input if the first and third digits are the same, and the middle is not equal to them, then it increases the second digit by three (if possible), otherwise it changes places the last two digits.